Remove dependency to korge-asset-manager and korge-parallax
Both korge addons were directly added into korge-fleks.

For korge-parallax this was done because only the parallax config
classes are used by korge-fleks. Also the config was adapted for
korge-fleks.

Code for korge-asset-manager was also moved back to korge-fleks because
of deeper integration with it. That means the standalone addon would not
work if not used with korge-fleks. This also removed dependencies
between those modules (korge-ldtk, korge-tiled).